requirements.txt
azure-devops==7.1.0b3
